moved the read profile function to utils
